Add files via upload
Now a day, we know that searching of jobs is so difficult in proficient areas the portal developed for the providing the simple and good job searching. With the help of this portal easily the job seeker can submit their resume and get the lot of opportunity of the job related to their profile. And by this website the companies or employer can also find the good and well profiled resume.
Now a day, we know that searching of jobs is so difficult in proficient areas the portal developed for the providing the simple and good job searching. With the help of this portal easily the job seeker can submit their resume and get the lot of opportunity of the job related to their profile. And by this website the companies or employer can also find the good and well profiled resume.
Now a day, we know that searching of jobs is so difficult in proficient areas the portal developed for the providing the simple and good job searching. With the help of this portal easily the job seeker can submit their resume and get the lot of opportunity of the job related to their profile. And by this website the companies or employer can also find the good and well profiled resume.

Software
•Database : SQLite
•Web Server : Internet Information Server
Technologies : HTML,CSS,,JavaScript,Bootstrap,python django